feat : 워크북 7

+++ sun010331/umc10th/src/main/java/com/example/umc10th/domain/review/dto/ReviewReqDTO.java
@@ -9,7 +9,7 @@ public class ReviewReqDTO {
             (
                     String title,
                     String content,
-                    BigDecimal star
+                    Float star
 
             ){
